# Rewrite `TypeEngine` for performance, robustness, and simplicity