--- a/sway-core/src/language/ty/declaration/storage.rs
+++ b/sway-core/src/language/ty/declaration/storage.rs
@@ -8,7 +8,7 @@ use sway_types::{Ident, Named, Span, Spanned};
 
 use crate::{
     engine_threading::*,
-    language::{parsed::StorageDeclaration, ty::*, Visibility},
+    language::{parsed::StorageDeclaration, ty::*},
     transform::{self},
     type_system::*,
     Namespace,
@@ -232,27 +232,6 @@ impl TyStorageDecl {
             },
             return_type,
         ))
-    }
-
-    pub(crate) fn fields_as_typed_struct_fields(&self) -> Vec<TyStructField> {
-        self.fields
-            .iter()
-            .map(
-                |TyStorageField {
-                     ref name,
-                     ref type_argument,
-                     ref span,
-                     ref attributes,
-                     ..
-                 }| TyStructField {
-                    visibility: Visibility::Public,
-                    name: name.clone(),
-                    span: span.clone(),
-                    type_argument: type_argument.clone(),
-                    attributes: attributes.clone(),
-                },
-            )
-            .collect()
     }
 }