--- a/sway-core/src/semantic_analysis/ast_node/code_block.rs
+++ b/sway-core/src/semantic_analysis/ast_node/code_block.rs
@@ -91,8 +91,6 @@ impl ty::TyCodeBlock {
         ctx: &TypeCheckContext,
         code_block: &TyCodeBlock,
     ) -> (TypeId, Span) {
-        let engines = ctx.engines();
-
         let implicit_return_span = code_block
             .contents
             .iter()
@@ -122,12 +120,8 @@ impl ty::TyCodeBlock {
                                 ..
                             }),
                         ..
-                    } => Some(
-                        ctx.engines
-                            .te()
-                            .insert(engines, TypeInfo::Never, span.source_id()),
-                    ),
-                    // find the implicit return, if any, and use it as the code block's return type.
+                    } => Some(ctx.engines.te().id_of_never()),
+                    // Find the implicit return, if any, and use it as the code block's return type.
                     // The fact that there is at most one implicit return is an invariant held by the parser.
                     ty::TyAstNode {
                         content:
@@ -153,11 +147,7 @@ impl ty::TyCodeBlock {
                     _ => None,
                 }
             })
-            .unwrap_or_else(|| {
-                ctx.engines
-                    .te()
-                    .insert(engines, TypeInfo::Tuple(Vec::new()), span.source_id())
-            });
+            .unwrap_or_else(|| ctx.engines.te().id_of_unit());
 
         (block_type, span)
     }