--- a/sway-core/src/semantic_analysis/ast_node/declaration/function/function_parameter.rs
+++ b/sway-core/src/semantic_analysis/ast_node/declaration/function/function_parameter.rs
@@ -1,7 +1,6 @@
 use crate::{
     language::{parsed::FunctionParameter, ty},
     semantic_analysis::{type_check_context::EnforceTypeArguments, TypeCheckContext},
-    type_system::*,
 };
 
 use sway_error::{
@@ -17,7 +16,6 @@ impl ty::TyFunctionParameter {
         parameter: FunctionParameter,
     ) -> Result<Self, ErrorEmitted> {
         let type_engine = ctx.engines.te();
-        let engines = ctx.engines();
 
         let FunctionParameter {
             name,
@@ -35,7 +33,7 @@ impl ty::TyFunctionParameter {
                 EnforceTypeArguments::Yes,
                 None,
             )
-            .unwrap_or_else(|err| type_engine.insert(engines, TypeInfo::ErrorRecovery(err), None));
+            .unwrap_or_else(|err| type_engine.id_of_error_recovery(err));
 
         type_argument.type_id.check_type_parameter_bounds(
             handler,
@@ -71,7 +69,6 @@ impl ty::TyFunctionParameter {
         parameter: &FunctionParameter,
     ) -> Result<Self, ErrorEmitted> {
         let type_engine = ctx.engines.te();
-        let engines = ctx.engines();
 
         let FunctionParameter {
             name,
@@ -90,7 +87,7 @@ impl ty::TyFunctionParameter {
                 EnforceTypeArguments::Yes,
                 None,
             )
-            .unwrap_or_else(|err| type_engine.insert(engines, TypeInfo::ErrorRecovery(err), None));
+            .unwrap_or_else(|err| type_engine.id_of_error_recovery(err));
 
         let typed_parameter = ty::TyFunctionParameter {
             name: name.clone(),