… sway-core/src/semantic_analysis/ast_node/declaration/struct.rs
… sway-core/src/semantic_analysis/ast_node/declaration/trait.rs
… sway-core/src/semantic_analysis/ast_node/declaration/trait_fn.rs
… sway-core/src/semantic_analysis/ast_node/declaration/trait_type.rs
… sway-core/src/semantic_analysis/ast_node/declaration/variable.rs
… sway-core/src/semantic_analysis/ast_node/expression/intrinsic_function.rs
… sway-core/src/semantic_analysis/ast_node/expression/match_expression/typed/instantiate.rs
… sway-core/src/semantic_analysis/ast_node/expression/match_expression/typed/matcher.rs
… sway-core/src/semantic_analysis/ast_node/expression/match_expression/typed/typed_match_branch.rs
… sway-core/src/semantic_analysis/ast_node/expression/match_expression/typed/typed_match_expression.rs
… sway-core/src/semantic_analysis/ast_node/expression/match_expression/typed/typed_scrutinee.rs
… sway-core/src/semantic_analysis/ast_node/expression/typed_expression.rs
… sway-core/src/semantic_analysis/ast_node/expression/typed_expression/enum_instantiation.rs
… sway-core/src/semantic_analysis/ast_node/expression/typed_expression/if_expression.rs
… sway-core/src/semantic_analysis/ast_node/expression/typed_expression/method_application.rs
… sway-core/src/semantic_analysis/ast_node/expression/typed_expression/struct_instantiation.rs
… sway-core/src/semantic_analysis/ast_node/mod.rs
… sway-core/src/semantic_analysis/namespace/trait_map.rs
… sway-core/src/semantic_analysis/node_dependencies.rs
… sway-core/src/semantic_analysis/type_check_context.rs
… sway-core/src/transform/to_parsed_lang/convert_parse_tree.rs
… sway-core/src/type_system/ast_elements/binding.rs
… sway-core/src/type_system/ast_elements/length.rs
… sway-core/src/type_system/ast_elements/trait_constraint.rs
… sway-core/src/type_system/ast_elements/type_argument.rs
… sway-core/src/type_system/ast_elements/type_parameter.rs
… sway-core/src/type_system/engine.rs
… sway-core/src/type_system/id.rs
… sway-core/src/type_system/info.rs
… sway-core/src/type_system/mod.rs