@ironcev 2024-10-06
While working on the improvements I've noticed an E2E test suddenly failing with this error message:
```
 7 │+  --> sway-lib-core/src/slice.sw:7:24
 8 │+   |
 9 │+ 5 |   impl<T> &__slice[T] {
10 │+ 6 |       pub fn ptr(self) -> raw_ptr {
11 │+ 7 |           let (ptr, _) = asm(s: self) {
12 │+   |  ________________________^
13 │+ 8 | |             s: (raw_ptr, u64)
14 │+ 9 | |         };
15 │+   | |_________^ Mismatched types.
16 │+expected: ()
17 │+found:    (pointer, u64).
18 │+help: Variable declaration's type annotation does not match up with the assigned expression's type.
19 │+10 |           ptr
20 │+11 |       }
```

Note that the wrongly reported compile error is in `core` library, means it should affect almost all the tests but it didn't. Indeed, re-running that test didn't reproduce the error.

I've kept re-running the E2E suit on another machine for two days straight (!!) before the error got provoked again, this time in another file:
```
   --> sway-lib-std/src/math.sw:316:28
    |
314 |           }
315 |   
316 |           let (a, b, c, d) = asm(r1: self) {
    |  ____________________________^
317 | |             r1: (u64, u64, u64, u64)
318 | |         };
    | |_________^ Mismatched types.
expected: ()
found:    (u64, u64, u64, u64).
help: Variable declaration's type annotation does not match up with the assigned expression's type.
319 |           if a != 0 {
320 |               return a.log2().as_u256() + 0xc0u256;
```

Up to now, I couldn't find a logical error in code, and chasing a non-reproducible issue is difficult indeed. But the issue is for sure still there. After opening this draft PR, the `forc-unit-tests` failed because of the same reason. It is green now, but the `cargo-run-e2e-test-release` failed on one test I am sure for the same reason.

It is interesting (and for sure good!), that the tests are failing on the build machine, but when run locally it takes literally tens of thousand of compilations to get the error.