@ironcev 2024-10-23
The changes in this PR make LSP garbage collection panic. A big thanks to @JoshuaBatty for detailed testing and finding this issue.

Joshua's tests showed that we already have GC issues on the master branch. Out of the 267 test projects in the `e2e_vm_tests/test_programs/should_pass/language` folder, we are getting 26 examples failing GC currently on the master branch. This PR makes additional six tests fail.

I'll keep the PR in draft until these issues are solved.