@ironcev 2024-11-04
All the GC issues reported in #6665, are now fixed on this branch. Similarly to the [transient bug explained above](https://github.com/FuelLabs/sway/pull/6613#issuecomment-2395539074), the root causes for GC related issues were already existing and the rewrite of the `TypeEngine` made them occur more frequently.

This PR now fixes, or at least mitigate to an extent, those GC issues which were possible to fix by _changing only the `TypeEngine` internals_.

However, GC still has issues whose solving requires restructurings outside of the `TypeEngine`. E.g. fixing #6687 requires changes in the `DeclEngine` as well.

The question is now how to proceed with this PR?

My thinking is to merge it as it is now, although knowing that not all of the GC issues are solved. The reasoning behind that is the following:
- this version already has more stable GC then the one on the `master`.
  When running GC tests on our whole E2E test suite (both `should_pass` and `should_fail` test) only one test fails in this PR (`should_pass/stdlib/option_eq/src/main.sw`), while numerous tests fail on `master`.

- the `option_eq` test fails on `master` as well, proving that this is not a new error caused by the changes in this branch.

- I've manually tried LSP from this PR on several representative projects end never experienced a crash. (Of course, `should_pass/stdlib/option_eq` project will cause LSP crash.)

- this PR is already vast in scope and brings significant changes. Extending it to rewrite the `DeclEngine` and address other GC related topics would be a huge scope-explosion. We originally planned to tackle the `DeclEngine` and "GC-friendliness" in separate PRs.

- keeping this PR on the top of other changes already proved to be a high effort. The PR touches the most prominent files on the frontend and is in constant collision with other changes done on the frontend.

@sdankel @JoshuaBatty @tritao @IGI-111 What do you think about this proposal? Can you also please try the local version of the LSP on your projects to see if you can experience any crashes?